Add functions to domUpdates.js
